feat: Add getSchemaNameApi function for retrieving schema name
The `getSchemaNameApi` function is added to the `GetSchemaNameApi.js` file in the `src/api/get` directory. This function makes a GET request to the server to retrieve the name of a schema based on the provided ID. The function utilizes the `getRequest` function from the existing "process" module to handle the API request.